refactor: move CheckUpdate to App class
