+++ app/version.properties
@@ -1,4 +1,4 @@
-#Sun Dec 09 19:02:05 CST 2018
+#Sun Dec 09 19:37:21 CST 2018
 VERSION_NUMBER=0
-VERSION_BUILD=158
+VERSION_BUILD=173
 VERSION_PATCH=0
